More changes in install script.

+++ install.sh
@@ -3,5 +3,7 @@ set -ex
 target=${target:-$HOME}
 cd config
 for file in *; do
-  ln -sf "$(pwd)/$file" $target/.$file
+  if [[ ! -e $target/.$file ]]; then
+	  ln -sf "$(pwd)/$file" $target/.$file
+  fi
 done
